+++ Genesis/src/utopia/genesis/shape/shape1D/Direction1D.scala
@@ -1,5 +1,6 @@
 package utopia.genesis.shape.shape1D
 
+import utopia.flow.util.RichComparable
 import utopia.genesis.shape.Axis2D
 import utopia.genesis.shape.shape2D.Direction2D
 import utopia.genesis.util.{Scalable, Signed}
@@ -9,7 +10,7 @@ import utopia.genesis.util.{Scalable, Signed}
   * @author Mikko Hilpinen
   * @since 17.4.2020, v2.3
   */
-sealed trait Direction1D
+sealed trait Direction1D extends RichComparable[Direction1D]
 {
 	// ABSTRACT	-----------------------------
 	
@@ -90,8 +91,13 @@ object Direction1D
 	case object Positive extends Direction1D
 	{
 		override val isPositive = true
-		
 		override val modifier = 1
+		
+		override def compareTo(o: Direction1D) = o match
+		{
+			case Positive => 0
+			case Negative => 1
+		}
 	}
 	
 	/**
@@ -100,8 +106,13 @@ object Direction1D
 	case object Negative extends Direction1D
 	{
 		override val isPositive = false
-		
 		override val modifier = -1
+		
+		override def compareTo(o: Direction1D) = o match
+		{
+			case Positive => -1
+			case Negative => 0
+		}
 	}
 	
 	/**